--- crates/iota-types/src/authentication/mod.rs
@@ -1,2 +1,0 @@
-pub mod auth_context;
-pub mod auth_digest;
