+++ crates/iota-types/src/error.rs
@@ -466,8 +466,8 @@ pub enum IotaError {
     InvalidAddress,
     #[error("Invalid transaction digest.")]
     InvalidTransactionDigest,
-    #[error("Invalid authentication digest.")]
-    InvalidAuthenticationDigest,
+    #[error("Invalid move authentication digest.")]
+    InvalidMoveAuthenticationDigest,
 
     #[error("Invalid digest length. Expected {expected}, got {actual}")]
     InvalidDigestLength { expected: usize, actual: usize },
