+++ crates/iota-types/src/lib.rs
@@ -32,7 +32,7 @@ use crate::{
 pub mod error;
 
 pub mod accumulator;
-pub mod authentication;
+pub mod auth_context;
 pub mod authenticator_state;
 pub mod balance;
 pub mod base_types;
